+++ src/domain/interfaces/IFileLocalStorage.ts
@@ -1,5 +1,0 @@
-import {IFileStorage} from './IFileStorage';
-
-export interface IFileLocalStorage extends IFileStorage {
-    getFilesPaths(): string[] | null;
-}