+++ src/domain/interfaces/IImagePreviewGeneratorResolver.ts
@@ -1,0 +1,8 @@
+import {FileModel} from '../models/FileModel';
+import {IImagePreviewGenerator} from './IImagePreviewGenerator';
+
+export const IImagePreviewGeneratorResolver = Symbol('IImagePreviewGeneratorResolver');
+
+export interface IImagePreviewGeneratorResolver {
+    resolve(file: FileModel): IImagePreviewGenerator,
+}