+++ src/infrastructure/tables/FileTable.ts
@@ -1,5 +1,7 @@
 import {TypeOrmTableFromModel} from '@steroidsjs/nest/infrastructure/decorators/typeorm/TypeOrmTableFromModel';
+import {Index} from 'typeorm';
 import {FileModel} from '../../domain/models/FileModel';
 
+@Index(['userId'])
 @TypeOrmTableFromModel(FileModel, 'file')
 export class FileTable extends FileModel {}